Split yEnc benchmarks into encoder and decoder classes
The combined YencBenchmarks held two [Benchmark(Baseline = true)] methods in
one class, which BenchmarkDotNet rejects (only one baseline allowed per class).
Split it into YencEncoderBenchmarks and YencDecoderBenchmarks, each with its
own baseline, and rename methods to the symmetric EncodeToLines/EncodeToWriter
and DecodeFromLines/DecodeFromBytes pairs.
